ansible/roles/acc-tableau-linux-three-node-ha/templates/content-backup/tableau-server-backup/requirements.txt
tableauserverclient==0.7
keyring==13.2.1
GitPython==2.1.11
gitdb2==3.0.1
